+++ src/com/company/Subtraction.java
@@ -1,9 +1,9 @@
 package com.company;
 
 public class Subtraction extends Operation{
-    int sub;
+    double sub;
 
-    public Subtraction(int number1, int number2){
+    public Subtraction(double number1, double number2){
 
         super( number1, number2, '-');
         this.sub = number1 - number2;
